+++ R/rMEA_class.R
@@ -3,13 +3,14 @@
 #' The preferred way to create an object of class \code{MEA} is through the function \code{\link{readMEA}}.
 #'
 #' @param dataframe a data frame with 2 columns containing MEA data respectively for subject 1 (s1) and subject 2 (s2).
-#' @param sampRate integer. The sampling rate of the MEA data. Normally derived from the framerate of the analyzed video sequence (frames per second; fps)
+#' @param sampRate integer. The sampling rate of the MEA data. Normally derived from the framerate of the analyzed video sequence (frames per second; fps).
 #' @param filter a string describing the pre-processing that has been applied on the raw data.
 #' @param id a string representing a unique identifier of the dyad that the MEA data belong to.
 #' @param session an integer representing the session (or experiment, interaction, etc); if each dyad is measured only once, the default value is 1.
 #' @param group a string naming the group the dyad belongs to, such as diagnostic group, clinic, etc.
-#' @param s1Name a string naming subject 1
-#' @param s2Name a string naming subject 2
+#' @param s1Name a string naming subject 1.
+#' @param s2Name a string naming subject 2.
+#' @param uid a string providing a unique identifier of the file. By default 'group_id_session'.
 #' @param x object to be tested.
 #' @details It is advised to \strong{not} create the \code{MEA} object manually but to always use the function \code{\link{readMEA}} instead.
 #'
@@ -23,8 +24,9 @@
 #'     ccfRes: some useful row marginals
 #' @export
 
-MEA = function(dataframe, sampRate, filter = "raw",id,
-               session, group, s1Name, s2Name
+MEA = function(dataframe, sampRate, filter = "raw", id,
+               session, group, s1Name, s2Name,
+               uid = paste(group,id,session,sep="_")
                ){
   x = list("MEA"=dataframe,
            "ccf" = NULL,
@@ -39,7 +41,7 @@ MEA = function(dataframe, sampRate, filter = "raw",id,
     ccf = "",
     s1Name = s1Name,
     s2Name = s2Name,
-    uid = paste(group,id,session,sep="_")
+    uid = uid
   ))
   class(x) = c("MEA",class(x))
   return(x)
@@ -55,7 +57,8 @@ MEA = function(dataframe, sampRate, filter = "raw",id,
 #' @return an object of class \code{MEAlist}
 #' @export
 MEAlist = function(listOfMea){
-  if(any(!sapply(listOfMea, is.MEA) ) ) stop("Only a list of MEA objects can be used to create a MEAlist", call.=F)
+  if(!is.list(listOfMea) || any(!sapply(listOfMea, is.MEA) ) ) stop("The supplied object must be a list containing only MEA objects", call.=F)
+  if(length(listOfMea)==0) stop("The supplied list is empty",call.=F)
   class(listOfMea) = "MEAlist"
   attributes(listOfMea) = c(attributes(listOfMea), list(
     nId = length(unique(sapply(listOfMea, attr, "id"))),
@@ -107,6 +110,13 @@ c.MEAlist = function(...){
   dots = lapply(dots, function(x){attributes(x)=NULL;x})
   MEAlist(do.call("c",dots))
 }
+
+#' @export
+"[.MEAlist" <- function(x,...,drop=FALSE){
+  y = NextMethod("[")
+  MEAlist(y)
+}
+
 
 
 #' Sets the group of MEA objects
@@ -209,4 +219,5 @@ summary.MEAlist = function(object, ...){
   invisible(Q)
 
 }
+